Fix various format string type mismatches.

+++ modules/chanserv/info.c
@@ -94,7 +94,7 @@ static void cs_cmd_info(sourceinfo_t *si, int parc, char *parv[])
 	if (CURRTIME - mc->used >= 86400)
 	{
 		if (hide_info)
-			command_success_nodata(si, _("Last used  : (about %d weeks ago)"), (CURRTIME - mc->used) / 604800);
+			command_success_nodata(si, _("Last used  : (about %d weeks ago)"), (int)((CURRTIME - mc->used) / 604800));
 		else
 		{
 			tm = *localtime(&mc->used);
